Renames two workflow-version components
Renames workflow-version-previous-detail-row to workflow-version-detail-row since it is for any workflow version.
Renames workflow-version-detail-row to workflow-latest-version-detail-row since it always shows the latest version.
