+++ app/components/workflow-latest-version-detail-row.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+
+const WorkflowLatestVersionDetailRow = Ember.Component.extend({
+  workflow: null
+});
+
+WorkflowLatestVersionDetailRow.reopenClass({
+  positionalParams: ['workflow']
+});
+
+export default WorkflowLatestVersionDetailRow;
