--- app/components/workflow-version-previous-detail-row.js
@@ -1,11 +1,0 @@
-import Ember from 'ember';
-
-const WorkflowVersionPreviousDetailRow = Ember.Component.extend({
-  workflow: null
-});
-
-WorkflowVersionPreviousDetailRow.reopenClass({
-  positionalParams: ['workflowVersion']
-});
-
-export default WorkflowVersionPreviousDetailRow;
