+++ app/templates/components/workflow-latest-version-detail-row.hbs
@@ -1,0 +1,9 @@
+<div class="workflow-latest-version-detail-row">
+  {{#workflow-version-summary workflow.latestVersion}}
+      <p class="small pull-right">{{#link-to "jobs.new.select-workflow-version" workflow}}
+          <span class="workflow-latest-version-detail-row-browse-versions">
+              Browse All Versions
+          </span>
+      {{/link-to}}</p>
+  {{/workflow-version-summary}}
+</div>
