+++ app/templates/components/workflow-version-detail-row.hbs
@@ -1,9 +1,3 @@
 <div class="workflow-version-detail-row">
-  {{#workflow-version-summary workflow.latestVersion}}
-      <p class="small pull-right">{{#link-to "jobs.new.select-workflow-version" workflow}}
-          <span class="workflow-version-detail-row-browse-versions">
-              Browse All Versions
-          </span>
-      {{/link-to}}</p>
-  {{/workflow-version-summary}}
+  {{workflow-version-summary workflowVersion}}
 </div>
