--- app/templates/components/workflow-version-previous-detail-row.hbs
@@ -1,3 +1,0 @@
-<div class="workflow-version-previous-detail-row">
-  {{workflow-version-summary workflowVersion}}
-</div>
