+++ app/templates/jobs/new/select-workflow-version.hbs
@@ -4,6 +4,6 @@
   title=title
   label="Workflow version"
   itemTitleKey="workflow.name"
-  detailComponent=(component 'workflow-version-previous-detail-row')}}
+  detailComponent=(component 'workflow-version-detail-row')}}
 {{/item-chooser}}
 {{outlet}}
