+++ app/templates/jobs/new/select-workflow.hbs
@@ -4,6 +4,6 @@
                 title="Please select a workflow to run"
                 label="Workflow"
                 itemTitleKey="name"
-                detailComponent=(component 'workflow-version-detail-row')}}
+                detailComponent=(component 'workflow-latest-version-detail-row')}}
 {{/item-chooser}}
 {{outlet}}
