+++ tests/integration/components/workflow-version-detail-row-test.js
@@ -5,15 +5,13 @@ moduleForComponent('workflow-version-detail-row', 'Integration | Component | wor
   integration: true
 });
 
-test('it renders workflow version details', function(assert) {
-  let workflow = {latestVersion: {version: 34, url: 'http://example.edu/workflows/workflow.cwl'} };
-  this.set('workflow', workflow);
-  this.render(hbs`{{workflow-version-detail-row workflow}}`);
+test('it renders', function(assert) {
 
-  //assert.equal(this.$().html(), 'taco');
-  assert.equal(this.$('.workflow-version-summary-version').text().trim(), '34');
-  assert.equal(this.$('.workflow-version-summary-url').attr('href'), 'http://example.edu/workflows/workflow.cwl');
+  let version = {version: 42, url: 'http://somwhere.edu/workflows/workflow.cwl'};
+  this.set('version', version);
+  this.render(hbs`{{workflow-version-detail-row version}}`);
+
+  assert.equal(this.$('.workflow-version-summary-version').text().trim(), '42');
+  assert.equal(this.$('.workflow-version-summary-url').attr('href'), 'http://somwhere.edu/workflows/workflow.cwl');
   assert.equal(this.$('.workflow-version-summary-description').text().trim(), 'Empty description');
-  assert.equal(this.$('.workflow-version-detail-row-browse-versions').text().trim(), 'Browse All Versions');
-
 });
